+++ orm/sveltekit/.gitignore
@@ -1,8 +1,24 @@
-.DS_Store
 node_modules
-/build
+/src/generated/prisma
+
+# Output
+.output
+.vercel
+.netlify
+.wrangler
 /.svelte-kit
-/package
+/build
+
+# OS
+.DS_Store
+Thumbs.db
+
+# Env
 .env
 .env.*
 !.env.example
+!.env.test
+
+# Vite
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
