+++ orm/sveltekit/package.json
@@ -1,39 +1,42 @@
 {
-  "name": "sveltekit",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
-  },
-  "devDependencies": {
-    "@sveltejs/adapter-auto": "4.0.0",
-    "@sveltejs/kit": "2.20.2",
-    "@sveltejs/package": "2.3.10",
-    "@typescript-eslint/eslint-plugin": "8.28.0",
-    "@typescript-eslint/parser": "8.28.0",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "10.1.1",
-    "eslint-plugin-svelte": "3.3.3",
-    "prettier": "3.5.3",
-    "prettier-plugin-svelte": "3.3.3",
-    "prisma": "6.5.0",
-    "svelte": "5.25.3",
-    "svelte-check": "4.1.5",
-    "svelte-preprocess": "6.0.3",
-    "tslib": "2.8.1",
-    "tsx": "4.19.3",
-    "typescript": "5.8.2",
-    "vite": "6.2.2"
-  },
-  "type": "module",
-  "dependencies": {
-    "@prisma/client": "6.5.0"
-  },
-  "prisma": {
-    "seed": "tsx prisma/seed.ts"
-  }
+	"name": "sveltekit2",
+	"private": true,
+	"version": "0.0.1",
+	"type": "module",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint ."
+	},
+	"prisma": {
+		"seed": "tsx prisma/seed.ts"
+	},
+	"devDependencies": {
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.16.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"eslint": "^9.18.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^3.0.0",
+		"globals": "^16.0.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prisma": "^6.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"tsx": "^4.19.3",
+		"typescript": "^5.0.0",
+		"typescript-eslint": "^8.20.0",
+		"vite": "^6.2.5"
+	},
+	"dependencies": {
+		"@prisma/client": "^6.6.0"
+	}
 }
