+++ orm/sveltekit/prisma/seed.ts
@@ -1,67 +1,44 @@
-import { PrismaClient, Prisma } from '@prisma/client'
+import { PrismaClient, Prisma } from '../src/generated/prisma';
 
-const prisma = new PrismaClient()
+const prisma = new PrismaClient();
 
 const userData: Prisma.UserCreateInput[] = [
-  {
-    name: 'Alice',
-    email: 'alice@prisma.io',
-    posts: {
-      create: [
-        {
-          title: 'Join the Prisma Discord',
-          content: 'https://pris.ly/discord',
-          published: true,
-        },
-      ],
-    },
-  },
-  {
-    name: 'Nilu',
-    email: 'nilu@prisma.io',
-    posts: {
-      create: [
-        {
-          title: 'Follow Prisma on Twitter',
-          content: 'https://www.twitter.com/prisma',
-          published: true,
-        },
-      ],
-    },
-  },
-  {
-    name: 'Mahmoud',
-    email: 'mahmoud@prisma.io',
-    posts: {
-      create: [
-        {
-          title: 'Ask a question about Prisma on GitHub',
-          content: 'https://www.github.com/prisma/prisma/discussions',
-          published: true,
-        },
-        {
-          title: 'Prisma on YouTube',
-          content: 'https://pris.ly/youtube',
-        },
-      ],
-    },
-  },
-]
+	{
+		name: 'Alice',
+		email: 'alice@prisma.io',
+		posts: {
+			create: [
+				{
+					title: 'Join the Prisma Discord',
+					content: 'https://pris.ly/discord',
+					published: true
+				},
+				{
+					title: 'Prisma on YouTube',
+					content: 'https://pris.ly/youtube'
+				}
+			]
+		}
+	},
+	{
+		name: 'Bob',
+		email: 'bob@prisma.io',
+		posts: {
+			create: [
+				{
+					title: 'Follow Prisma on Twitter',
+					content: 'https://www.twitter.com/prisma',
+					published: true
+				}
+			]
+		}
+	}
+];
 
-async function main() {
-  for (const u of userData) {
-    const user = await prisma.user.create({
-      data: u,
-    })
-  }
+export async function main() {
+	for (const u of userData) {
+		await prisma.user.create({ data: u });
+	}
 }
 
-main()
-  .then(async () => {
-    await prisma.$disconnect()
-  })
-  .catch(async (e) => {
-    console.error(e)
-    await prisma.$disconnect()
-    process.exit(1)
-  })
+main();
