--- orm/sveltekit/src/routes/drafts/+page.server.ts
@@ -1,11 +1,0 @@
-import prisma from '$lib/prisma';
-
-/** @type {import('./$types').PageServerLoad} */
-export const load = async () => {
-  const response = await prisma.post.findMany({
-    where: { published: false },
-    include: { author: true },
-  })
-
-  return { drafts: response };
-};
