--- orm/sveltekit/vite.config.js
@@ -1,8 +1,0 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-
-/** @type {import('vite').UserConfig} */
-const config = {
-  plugins: [sveltekit()]
-};
-
-export default config;
