--- orm/sveltekit2/src/lib/prisma.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from '../generated/prisma';
-
-const prisma = new PrismaClient();
-
-export default prisma;
