--- orm/sveltekit2/src/routes/+page.server.ts
@@ -1,13 +1,0 @@
-import prisma from '$lib/prisma';
-
-export async function load() {
-	const users = await prisma.user.findMany({
-		include: {
-			posts: true
-		}
-	});
-
-	return {
-		users
-	};
-}
